pyproject.toml
[project]
name = "dev-smart"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref==3.9.1",
    "attrs==25.3.0",
    "autobahn==24.4.2",
    "automat==25.4.16",
    "cffi==1.17.1",
    "channels==4.2.2",
    "channels-redis==4.2.1",
    "colorama==0.4.6",
    "constantly==23.10.4",
    "cryptography==45.0.5",
    "daphne==4.2.1",
    "django==5.2.4",
    "environs==14.2.0",
    "hyperlink==21.0.0",
    "idna==3.10",
    "incremental==24.7.2",
    "iniconfig==2.1.0",
    "marshmallow==4.0.0",
    "msgpack==1.1.1",
    "openpyxl==3.1.2",
    "packaging==25.0",
    "pillow==11.3.0",
    "pluggy==1.6.0",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pygments==2.19.2",
    "pyinstaller==6.16.0",
    "pyopenssl==25.1.0",
    "pyserial==3.5",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "python-dotenv==1.1.1",
    "pytz==2025.2",
    "pyusb==1.2.1",
    "redis==6.2.0",
    "requests>=2.32.5",
    "ruff==0.12.4",
    "service-identity==24.2.0",
    "setuptools==80.9.0",
    "sqlparse==0.5.3",
    "twisted==25.5.0",
    "txaio==25.6.1",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "zope-interface==7.2",
]
